+++ apps/web/src/features/moim-edit/ui/moim-edit-form.tsx
@@ -1,8 +1,7 @@
 "use client";
 
 import { useRouter } from "next/navigation";
-import { useState } from "react";
-import { uploadImage } from "@/_pages/moim-create/use-cases/upload-image";
+import { useMoimFormImageUpload } from "@/features/moim-create/hooks/use-moim-form-image-upload";
 import type { MoimCreateFormValues } from "@/features/moim-create/model/schema";
 import { MoimFormFields } from "@/features/moim-create/ui/moim-form-fields";
 import { useMoimEditForm } from "@/features/moim-edit/model/use-moim-edit-form";
@@ -14,40 +13,12 @@ interface MoimEditFormProps {
 
 export const MoimEditForm = ({ meetingId, initialValues }: MoimEditFormProps) => {
   const router = useRouter();
-  const [isImageUploading, setIsImageUploading] = useState(false);
   const { form, onSubmit, state, isPending } = useMoimEditForm({
     meetingId,
     initialValues,
   });
 
-  const { setError, clearErrors } = form;
-
-  const handleImageUpload = async (onChange: (url: string) => void) => {
-    const input = document.createElement("input");
-    input.type = "file";
-    input.accept = "image/jpeg,image/png,image/webp,image/gif";
-
-    input.onchange = async (event) => {
-      const file = (event.target as HTMLInputElement).files?.[0];
-      if (!file) return;
-
-      setIsImageUploading(true);
-      try {
-        const publicUrl = await uploadImage(file);
-        onChange(publicUrl);
-        clearErrors("image");
-      } catch {
-        setError("image", {
-          type: "manual",
-          message: "이미지 업로드에 실패했습니다. 다시 시도해주세요.",
-        });
-      } finally {
-        setIsImageUploading(false);
-      }
-    };
-
-    input.click();
-  };
+  const { isImageUploading, handleImageUpload } = useMoimFormImageUpload(form);
 
   return (
     <MoimFormFields
